+++ test/test_ascii.py
@@ -9,8 +9,8 @@ _viscid_root = os.path.realpath(os.path.dirname(__file__) + '/../viscid/')
 if not _viscid_root in sys.path:
     sys.path.append(_viscid_root)
 
+import viscid
 from viscid import vutil
-from viscid import readers
 from viscid.plot import mpl
 
 def main():
@@ -18,7 +18,7 @@ def main():
     parser.add_argument("--show", "--plot", action="store_true")
     args = vutil.common_argparse(parser)
 
-    f = readers.load_file(_viscid_root + '/../sample/test.asc')
+    f = viscid.load_file(_viscid_root + '/../sample/test.asc')
     mpl.plot(f['c1'], show=args.show)
 
 if __name__ == "__main__":