+++ viscid/readers/ggcm_fortbin.py
@@ -254,7 +254,7 @@ class GGCMFortbinFileWrapper(object):
         self.open()
         return self
 
-    def __exit__(self, typ, value, traceback):
+    def __exit__(self, exc_type, value, traceback):
         self.close()
 
     def _read_file_header(self, data_size=4):