+++ viscid/readers/xdmf.py
@@ -410,8 +410,8 @@ class FileXDMF(vfile.VFile):  # pylint: disable=abstract-method
         data, dataattrs = self._parse_dataitem(item.find("./DataItem"))
         name = attrs["Name"]
         center = attrs["Center"]
-        typ = attrs["AttributeType"]
-        fld = self._make_field(parent_node, typ, name, crds, data,
+        fldtype = attrs["AttributeType"]
+        fld = self._make_field(parent_node, fldtype, name, crds, data,
                                center=center, time=time)
         return fld