+++ CHANGES.md
@@ -2,7 +2,13 @@
 
 ## 0.80.7 dev
 
+Backward Incompatible Changes:
+  - Arguments to viscid.field.wrap_field have changed
+  - typ arguments have been refactored to fldtype and crdtype
 
+Changes:
+  - add viscid.field.arrays2field and viscid.coordinates.arrays2crds
+  - viscid.field now has empty, ones, zeros, empty_like, ones_like, zeros_like that act roughly the same as the numpy functions, except that these create fields
 
 ## 0.80.6